feat: restructure media details ui components
