+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/feature/details/media/ui/components/TitleDetailsScreenshots.kt
@@ -1,0 +1,18 @@
+package com.divinelink.scenepeek.feature.details.media.ui.components
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import com.divinelink.feature.details.media.ui.components.TitleDetailsPreview
+import com.divinelink.feature.details.media.ui.components.TitleTvDetailsPreview
+
+@Preview
+@Composable
+fun TitleDetailsScreenshots() {
+  TitleDetailsPreview()
+}
+
+@Preview
+@Composable
+fun TitleTvDetailsScreenshots() {
+  TitleTvDetailsPreview()
+}
